code splite by components
