+++ package.json
@@ -2,7 +2,7 @@
   "name": "ql-platform-ui",
   "version": "0.2.6",
   "description": "react component",
-  "main": "lib/index.js",
+  "main": "lib/index/index.js",
   "scripts": {
     "test": "jest",
     "example": "webpack --config build/webpack.demo.config.js",
@@ -39,10 +39,12 @@
     "babel-loader": "^8.0.5",
     "clean-webpack-plugin": "^1.0.1",
     "css-loader": "^2.1.0",
+    "glob": "^7.1.4",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.8.0",
     "less": "^3.9.0",
     "less-loader": "^4.1.0",
+    "mini-css-extract-plugin": "^0.7.0",
     "react-test-render": "^1.1.1",
     "style-loader": "^0.23.1",
     "typescript": "^3.4.5",
@@ -54,7 +56,6 @@
     "classnames": "^2.2.6",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.12.1",
-    "glob": "^7.1.3",
     "popper.js": "^1.15.0",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
