+++ src/components/Breadcrumb/index.tsx
@@ -1,3 +1,6 @@
 import Breadcrumb from './Breadcrumb'
 
+export {
+    Breadcrumb
+}
 export default Breadcrumb
